+++ src/main/java/it/lbsoftware/daily/tags/TagController.java
@@ -6,6 +6,7 @@ import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
+import org.springframework.web.server.ResponseStatusException;
 
 import javax.validation.Valid;
 import java.util.List;
@@ -26,7 +27,11 @@ class TagController {
 
     @PostMapping
     public ResponseEntity<TagDto> createTag(@Valid @RequestBody TagDto tagDto) {
-        AppUser appUser = appUserService.getAppUserFromToken();
+        Optional<AppUser> appUserOptional = appUserService.getAppUserFromToken();
+        if(appUserOptional.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User could not be found");
+        }
+        AppUser appUser = appUserOptional.get();
         Tag tag = tagDtoMapper.convertToEntity(tagDto);
         Tag createdTag = tagService.createTag(tag, appUser);
         TagDto createdTagDto = tagDtoMapper.convertToDto(createdTag);
@@ -36,7 +41,11 @@ class TagController {
 
     @GetMapping(value = "/{uuid}")
     public ResponseEntity<TagDto> readTag(@PathVariable("uuid") UUID uuid) {
-        AppUser appUser = appUserService.getAppUserFromToken();
+        Optional<AppUser> appUserOptional = appUserService.getAppUserFromToken();
+        if(appUserOptional.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User could not be found");
+        }
+        AppUser appUser = appUserOptional.get();
         Optional<Tag> readTag = tagService.readTag(uuid, appUser);
         if (readTag.isEmpty()) {
             return ResponseEntity.notFound().build();
@@ -48,7 +57,11 @@ class TagController {
 
     @GetMapping
     public ResponseEntity<List<TagDto>> readTags() {
-        AppUser appUser = appUserService.getAppUserFromToken();
+        Optional<AppUser> appUserOptional = appUserService.getAppUserFromToken();
+        if(appUserOptional.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User could not be found");
+        }
+        AppUser appUser = appUserOptional.get();
         List<Tag> readTags = tagService.readTags(appUser);
         List<TagDto> readTagDtos = readTags.stream().map(tagDtoMapper::convertToDto).collect(Collectors.toList());
 
@@ -57,7 +70,11 @@ class TagController {
 
     @PutMapping(value = "/{uuid}")
     public ResponseEntity<TagDto> updateTag(@PathVariable("uuid") UUID uuid, @Valid @RequestBody TagDto tagDto) {
-        AppUser appUser = appUserService.getAppUserFromToken();
+        Optional<AppUser> appUserOptional = appUserService.getAppUserFromToken();
+        if(appUserOptional.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User could not be found");
+        }
+        AppUser appUser = appUserOptional.get();
         Tag tag = tagDtoMapper.convertToEntity(tagDto);
         Optional<Tag> updatedTag = tagService.updateTag(uuid, tag, appUser);
         if (updatedTag.isEmpty()) {
@@ -69,7 +86,11 @@ class TagController {
 
     @DeleteMapping(value = "/{uuid}")
     public ResponseEntity<TagDto> deleteTag(@PathVariable("uuid") UUID uuid) {
-        AppUser appUser = appUserService.getAppUserFromToken();
+        Optional<AppUser> appUserOptional = appUserService.getAppUserFromToken();
+        if(appUserOptional.isEmpty()){
+            throw new ResponseStatusException(HttpStatus.UNAUTHORIZED, "User could not be found");
+        }
+        AppUser appUser = appUserOptional.get();
         if (!tagService.deleteTag(uuid, appUser)) {
             return ResponseEntity.notFound().build();
         }
